// Iverson.Clients/TypeScript/generated/object_retrieval.ts
export interface RetrievalManyRequest {
  typeName: string;
  keys: string[];
  traceId: string;
}

export interface RetrievalResponse {
  found: boolean;
  data: { [key: string]: any } | undefined;
  traceId: string;
}
// ...
export const RetrievalManyRequest: MessageFns<RetrievalManyRequest> = {
  encode(message: RetrievalManyRequest, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.typeName !== "") {
      writer.uint32(10).string(message.typeName);
    }
    for (const v of message.keys) {
      writer.uint32(18).string(v!);
    }
    if (message.traceId !== "") {
      writer.uint32(26).string(message.traceId);
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): RetrievalManyRequest {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseRetrievalManyRequest();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.typeName = reader.string();
          continue;
        }
        case 2: {
          if (tag !== 18) {
            break;
          }

          message.keys.push(reader.string());
          continue;
        }
        case 3: {
          if (tag !== 26) {
            break;
          }

          message.traceId = reader.string();
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): RetrievalManyRequest {
    return {
      typeName: isSet(object.typeName)
        ? globalThis.String(object.typeName)
        : isSet(object.type_name)
        ? globalThis.String(object.type_name)
        : "",
      keys: globalThis.Array.isArray(object?.keys) ? object.keys.map((e: any) => globalThis.String(e)) : [],
      traceId: isSet(object.traceId)
        ? globalThis.String(object.traceId)
        : isSet(object.trace_id)
        ? globalThis.String(object.trace_id)
        : "",
    };
  },

  toJSON(message: RetrievalManyRequest): unknown {
    const obj: any = {};
    if (message.typeName !== "") {
      obj.typeName = message.typeName;
    }
    if (message.keys?.length) {
      obj.keys = message.keys;
    }
    if (message.traceId !== "") {
      obj.traceId = message.traceId;
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<RetrievalManyRequest>, I>>(base?: I): RetrievalManyRequest {
    return RetrievalManyRequest.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<RetrievalManyRequest>, I>>(object: I): RetrievalManyRequest {
    const message = createBaseRetrievalManyRequest();
    message.typeName = object.typeName ?? "";
    message.keys = object.keys?.map((e) => e) || [];
    message.traceId = object.traceId ?? "";
    return message;
  },
};

function createBaseRetrievalResponse(): RetrievalResponse {
  return { found: false, data: undefined, traceId: "" };
}

export const RetrievalResponse: MessageFns<RetrievalResponse> = {
  encode(message: RetrievalResponse, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.found !== false) {
      writer.uint32(8).bool(message.found);
    }
    if (message.data !== undefined) {
      Struct.encode(Struct.wrap(message.data), writer.uint32(18).fork()).join();
    }
    if (message.traceId !== "") {
      writer.uint32(26).string(message.traceId);
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): RetrievalResponse {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseRetrievalResponse();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 8) {
            break;
          }

          message.found = reader.bool();
          continue;
        }
        case 2: {
          if (tag !== 18) {
            break;
          }

          message.data = Struct.unwrap(Struct.decode(reader, reader.uint32()));
          continue;
        }
        case 3: {
          if (tag !== 26) {
            break;
          }

          message.traceId = reader.string();
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): RetrievalResponse {
    return {
      found: isSet(object.found) ? globalThis.Boolean(object.found) : false,
      data: isObject(object.data) ? object.data : undefined,
      traceId: isSet(object.traceId)
        ? globalThis.String(object.traceId)
        : isSet(object.trace_id)
        ? globalThis.String(object.trace_id)
        : "",
    };
  },

  toJSON(message: RetrievalResponse): unknown {
    const obj: any = {};
    if (message.found !== false) {
      obj.found = message.found;
    }
    if (message.data !== undefined) {
      obj.data = message.data;
    }
    if (message.traceId !== "") {
      obj.traceId = message.traceId;
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<RetrievalResponse>, I>>(base?: I): RetrievalResponse {
    return RetrievalResponse.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<RetrievalResponse>, I>>(object: I): RetrievalResponse {
    const message = createBaseRetrievalResponse();
    message.found = object.found ?? false;
    message.data = object.data ?? undefined;
    message.traceId = object.traceId ?? "";
    return message;
  },
};
// ...
type Builtin = Date | Function | Uint8Array | string | number | boolean | undefined;

export type DeepPartial<T> = T extends Builtin ? T
  : T extends globalThis.Array<infer U> ? globalThis.Array<DeepPartial<U>>
  : T extends ReadonlyArray<infer U> ? ReadonlyArray<DeepPartial<U>>
  : T extends {} ? { [K in keyof T]?: DeepPartial<T[K]> }
  : Partial<T>;

type KeysOfUnion<T> = T extends T ? keyof T : never;
export type Exact<P, I extends P> = P extends Builtin ? P
  : P & { [K in keyof P]: Exact<P[K], I[K]> } & { [K in Exclude<keyof I, KeysOfUnion<P>>]: never };

function isObject(value: any): boolean {
  return typeof value === "object" && value !== null;
}

function isSet(value: any): boolean {
  return value !== null && value !== undefined;
}

export interface MessageFns<T> {
  encode(message: T, writer?: BinaryWriter): BinaryWriter;
  decode(input: BinaryReader | Uint8Array, length?: number): T;
  fromJSON(object: any): T;
  toJSON(message: T): unknown;
  create<I extends Exact<DeepPartial<T>, I>>(base?: I): T;
  fromPartial<I extends Exact<DeepPartial<T>, I>>(object: I): T;
}